feat: Update version to 2.3.1 and implement custom customer creation logic
